ModelTraningAndTesting/requirements.txt
torch
torchvision
torchaudio
foolbox
eagerpy
opencv-python
tensorboard
torchinfo
tqdm
numpy
pillow